Correctifs recover_dist_tree

See merge request numerics/mesh/maia!149

+++ maia/transfer/utils.py
@@ -70,7 +70,7 @@ def create_all_elt_g_numbering(p_zone, dist_elts):
   sorting_idx = np.argsort([PT.Element.Range(elt)[0] for elt in dist_elts])
   sorted_dist_elts  = [dist_elts[k] for k in sorting_idx]
   elt_sections_dn   = [PT.Element.Size(elt) for elt in sorted_dist_elts]
-  elt_sections_idx  = np_utils.sizes_to_indices(elt_sections_dn, dtype=np.int32)
+  elt_sections_idx  = np_utils.sizes_to_indices(elt_sections_dn, dtype=pdm_gnum_dtype)
   p_elts = [PT.get_node_from_name(p_zone, PT.get_name(elt)) for elt in sorted_dist_elts]
   elt_sections_pn = [MT.getGlobalNumbering(elt, 'Element')[1].size if elt else 0 for elt in p_elts]
   offset = 0
